requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
streamlit==1.49.1
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
httpx==0.28.1
requests==2.32.5

langchain==0.3.27
langchain-community==0.3.29
langchain-text-splitters==0.3.11
langchain-ollama==0.3.8

faiss-cpu==1.12.0
sentence-transformers==5.1.0

pymupdf==1.26.4
python-docx==1.2.0

ragas==0.3.4
datasets==4.0.0
pandas==2.3.2

pytest==8.4.1
